Rename: "폴더추가 및 경로 변경, js/renderer.js -> view/renderer.js"
